+++ PavlovRconWebserver/Controllers/SshServerController.cs
@@ -2,16 +2,17 @@ using System;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using LiteDB;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using PavlovRconWebserver.Exceptions;
+using PavlovRconWebserver.Extensions;
 using PavlovRconWebserver.Models;
 using PavlovRconWebserver.Services;
 
 namespace PavlovRconWebserver.Controllers
 {
-    
-    [Authorize(Roles = CustomRoles.Admin)]
+    [Authorize(Roles = CustomRoles.OnPremiseOrRent)]
     public class SshServerController : Controller
     {
         private readonly MapsService _mapsService;
@@ -34,52 +35,58 @@ namespace PavlovRconWebserver.Controllers
         [HttpGet("[controller]/")]
         public async Task<IActionResult> Index()
         {
-            return View("Index", await _service.FindAll());
+            var user = await _userservice.getUserFromCp(HttpContext.User);
+            return View("Index", await _service.FindAllWithRightsCheck(HttpContext.User, user));
         }
 
+        [Authorize(Roles = CustomRoles.OnPremise)]
         [HttpGet("[controller]/EditServer/{serverId?}")]
         public async Task<IActionResult> EditServer(int? serverId)
         {
+            if (serverId!= null && !await CheckIfUserHasrights((int) serverId)) return Forbid();
             var server = new SshServer();
             if (serverId != null && serverId != 0) server = await _service.FindOne((int) serverId);
-
-            try
-            {
-                server.SshKeyFileNames = Directory.EnumerateFiles("KeyFiles/", "*", SearchOption.AllDirectories)
-                    .Select(x => x.Replace("KeyFiles/", "")).ToList();
-            }
-            catch (Exception)
-            {
-                // ignore there is maybe no folder or the folder is empty 
-            }
-
-            ViewBag.Remove = false;
+            if(server.Id==0)
+                if (!HttpContext.User.IsInRole("Admin"))
+                    return Forbid();
+            ViewBag.Remove = false;;
+            server.LiteDbUsers = (await _userservice.FindAll()).ToList();
             return View("Server", server);
         }
 
+        [Authorize(Roles = CustomRoles.OnPremise)]
         [HttpPost]
         public async Task<IActionResult> EditServer(SshServer server, bool remove = false)
         {
-            try
-            {
-                server.SshKeyFileNames = Directory.EnumerateFiles("KeyFiles/", "*", SearchOption.AllDirectories)
-                    .Select(x => x.Replace("KeyFiles/", "")).ToList();
-            }
-            catch (Exception)
-            {
-                // ignore there is maybe no folder or the folder is empty 
-            }
+            if (!await CheckIfUserHasrights(server.Id)) return Forbid();
 
             ViewBag.Remove = remove;
+            
+            server.LiteDbUsers = (await _userservice.FindAll()).ToList();
             return View("Server", server);
         }
 
+        [Authorize(Roles = CustomRoles.OnPremise)]
         [HttpPost("[controller]/SaveServer")]
         public async Task<IActionResult> SaveServer(SshServer server)
         {
+            if (!await CheckIfUserHasrights(server.Id)) return Forbid();
             var newServer = false;
             if (!ModelState.IsValid)
                 return View("Server", server);
+
+            server.LiteDbUsers = (await _userservice.FindAll()).ToList();
+            if (server.SshKeyFileNameForm != null)
+            {
+                await using var ms = new MemoryStream();
+                await server.SshKeyFileNameForm.CopyToAsync(ms);
+                var fileBytes = ms.ToArray();
+                server.SshKeyFileName = fileBytes;
+                // act on the Base64 data
+            }
+            
+            server.Owner = (await _userservice.FindAll())
+                .FirstOrDefault(x => x.Id == new ObjectId(server.LiteDbUserId));
             try
             {
                 if (server.Id == 0)
@@ -89,6 +96,14 @@ namespace PavlovRconWebserver.Controllers
                 }
                 else
                 {
+                    if (server.SshKeyFileName == null)
+                    {
+                        var old = await _service.FindOne(server.Id);
+                        if (old.SshKeyFileName != null)
+                        {
+                            server.SshKeyFileName = old.SshKeyFileName;
+                        }
+                    }
                     await _service.Update(server);
                 }
             }
@@ -120,6 +135,9 @@ namespace PavlovRconWebserver.Controllers
         [HttpGet("[controller]/SaveServerSelectedMap")]
         public async Task<bool> SaveServerSelectedMap(int serverId, string mapId, string gameMode)
         {
+            if (!await RightsHandler.HasRightsToThisPavlovServer(HttpContext.User,
+                await _userservice.getUserFromCp(HttpContext.User), serverId, _service, _pavlovServerService))
+                return false;
             var realMap = await _mapsService.FindOne(mapId);
             var pavlovServer = await _pavlovServerService.FindOne(serverId);
             var mapsSelected = await _serverSelectedMapService.FindAllFrom(pavlovServer);
@@ -158,7 +176,10 @@ namespace PavlovRconWebserver.Controllers
 
         [HttpGet("[controller]/DeleteServerSelectedMap")]
         public async Task<bool> DeleteServerSelectedMap(int serverId, string mapId, string gameMode)
-        {
+        {  
+            if (!await RightsHandler.HasRightsToThisPavlovServer(HttpContext.User,
+                await _userservice.getUserFromCp(HttpContext.User), serverId, _service, _pavlovServerService))
+                return false;
             var realMap = await _mapsService.FindOne(mapId);
             var pavlovServer = await _pavlovServerService.FindOne(serverId);
             var mapsSelected = await _serverSelectedMapService.FindAllFrom(pavlovServer);
@@ -175,8 +196,17 @@ namespace PavlovRconWebserver.Controllers
         [HttpGet("[controller]/DeleteServer/{id}")]
         public async Task<IActionResult> DeleteServer(int id)
         {
+            if (!await CheckIfUserHasrights(id)) return Forbid();
             await _service.Delete(id);
             return await Index();
+        }
+
+        private async Task<bool> CheckIfUserHasrights(int id)
+        {
+            var user = await _userservice.getUserFromCp(HttpContext.User);
+            if (await RightsHandler.HasRightsToThisSshServer(HttpContext.User, user, id, _service)) return true;
+
+            return false;
         }
     }
 }
